requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==20.1.0
idna==3.7
inflection==0.5.1
lxml==5.3.0
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==0.21.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==73.0.1
soupsieve==2.6
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.2.0
